chore: rename files with types (step 2/2)
