+++ src/resolvers/__tests__/updateOne-test.ts
@@ -4,7 +4,7 @@ import { Resolver, schemaComposer, ObjectTypeComposer } from 'graphql-compose';
 import { GraphQLNonNull } from 'graphql-compose/lib/graphql';
 import { UserModel, IUser } from '../../__mocks__/userModel';
 import updateOne from '../updateOne';
-import GraphQLMongoID from '../../types/mongoid';
+import GraphQLMongoID from '../../types/MongoID';
 import { convertModelToGraphQL } from '../../fieldsConverter';
 import { ExtendedResolveParams } from '..';
